conclusione controlli

+++ back-end/class/query/news.php
@@ -21,9 +21,19 @@ class news extends connection implements query{
         $this->regex=regex::date();
     }
     public function write(){
-        preg_match($this->regex, $this->date, $matches, PREG_OFFSET_CAPTURE);
-        if(empty($matches)){new exeption('error','Orario non in formato GG-MM-AAAA.');}
-
+        if($this->type == 'Cerca nel menu:'){
+            throw new exeption('error',"non è stato definito il tipo dell'evento");
+        }
+        if(!preg_match($this->regex, $this->date, $matches, PREG_OFFSET_CAPTURE)){
+            throw new exeption('error','Orario non in formato GG-MM-AAAA.');
+        }
+        if(strlen($this->news) > 64){
+            throw new exeption('error','lo novità inserita deve essere di lunghezza minore di 64 caratteri');
+        }
+        if($this->news == 'Inserisci la nuova novità'){
+            throw new exeption('error','non è stata inserita alcuna novità');
+        }
+        
         $date_array = explode("-",$this->date); // split the array
         $var_day = $date_array[0]; //day seqment
         $var_month = $date_array[1]; //month segment
@@ -32,7 +42,7 @@ class news extends connection implements query{
 
         $query = "INSERT INTO eventi VALUES (NULL,'$this->type','$new_date_format','$this->news')";
         if(!parent::execute_query($query)){
-            throw new exeption("error","Non è stato possibile inserire la novita. Riprovare più tardi.");
+            throw new exeption("error","Non è stato possibile inserire la novità. Riprovare più tardi.");
         }
     }
     public function read($type=NULL){
@@ -43,6 +53,9 @@ class news extends connection implements query{
     }
 
     public function delete(){
+        if($this->type == 'nessuno'){
+            throw new exeption('error',"non è stato definito il tipo dell'evento");
+        }
         $query="DELETE FROM eventi WHERE ID = '$this->type'";
         if(!parent::execute_query($query)){
             throw new exeption('error', 'evento con id '.$this->id.' non eliminato. Riprovare più tardi.');
